+++ ldd.php
@@ -70,6 +70,7 @@ if ( ! ldd_requirements_check() ) {
 	return false;
 }
 
+require_once LDD_DIR_INC . 'functions.php';
 require_once LDD_DIR_INC . 'class-ldd.php';